requirements.txt
streamlit
ffmpeg-python
librosa
numpy
pandas
scikit-learn
tensorflow
keras
ipython
